requirements.txt
# Core dependencies
requests==2.31.0          # HTTP 请求库
pyyaml==6.0.1            # YAML 配置支持
rich==13.7.0             # 终端美化
boto3>=1.34.0            # AWS SDK

# Async support (optional)
aiohttp>=3.8.0           # 异步 HTTP

# Logging
python-json-logger==2.0.7 # JSON 格式日志

# Development
pytest==7.4.3            # 单元测试
black==23.11.0           # 代码格式化

# 重试机制（目前已注释相关代码，可选）
# tenacity>=8.2.3

# 测试相关（可选）
# pytest>=7.4.4
# pytest-asyncio>=0.23.3

# 其他可能需要的包
python-dateutil>=2.8.2
requests-toolbelt>=1.0.0